* extracted customLogger from logTest and kept separate, so we can validate any log messages * validated the warning log when there is a wrong doc-id is passed and resolves
